[API] Use `int32` instead of `int8/16` (#1074)

+++ tests/test_othello.py
@@ -25,7 +25,7 @@ def test_step():
     state = step(state, 34)
     state = step(state, 17)
     # fmt: off
-    expected = jnp.int8([
+    expected = jnp.int32([
         0, 0, 0, 0, 0, 0, 0, 0,
         0, 0, 0, 0, 0, 0, 0, 0,
         0, -1, -1, -1, -1, -1, 0, 0,
@@ -120,12 +120,12 @@ def test_random_play():
     done = jnp.bool_(False)
     key, sub_key = jax.random.split(key)
     state = init(sub_key)
-    rewards = jnp.int16([0.0, 0.0])
+    rewards = jnp.int32([0.0, 0.0])
     while not done:
         legal_actions = jnp.where(state.legal_action_mask)[0]
         key, sub_key = jax.random.split(key)
         action = jax.random.choice(sub_key, legal_actions)
-        state = step(state, jnp.int16(action))
+        state = step(state, jnp.int32(action))
         done = state.terminated
         rewards += state.rewards
